requirements.txt
jyboss
PyYAML>=3.0
simplejson >= 3.1